+++ src/Api/Request.php
@@ -57,7 +57,7 @@ class Request
         'DNT: 1',
         'X-Pinterest-AppState: active',
         'X-NEW-APP: 1',
-        'X-APP-VERSION: f2d280f',
+        'X-APP-VERSION: 2414ec2',
         'X-Requested-With: XMLHttpRequest',
     ];