+++ src/Helpers/UrlBuilder.php
@@ -52,6 +52,9 @@ class UrlBuilder
     const RESOURCE_USER_FEED = 'resource/UserHomefeedResource/get/';
     const RESOURCE_RELATED_PINS = 'resource/RelatedPinFeedResource/get';
     const RESOURCE_VISUAL_SIMILAR_PINS = 'resource/VisualLiveSearchResource/get';
+    const RESOURCE_BULK_COPY = 'resource/BulkEditResource/create/';
+    const RESOURCE_BULK_MOVE = 'resource/BulkEditResource/update/';
+    const RESOURCE_BULK_DELETE = 'resource/BulkEditResource/delete/';
 
     /**
      * Pinners